:hammer: Refactor SyncManager and SyncHandler for multi-platform reuse in iOS and Android
