+++ composeApp/src/commonMain/kotlin/com/crosspaste/signal/PreKeyBundleCodecs.kt
@@ -1,0 +1,8 @@
+package com.crosspaste.signal
+
+interface PreKeyBundleCodecs {
+
+    fun decodePreKeyBundle(encoded: ByteArray): PreKeyBundleInterface
+
+    fun encodePreKeyBundle(preKeyBundleInterface: PreKeyBundleInterface): ByteArray
+}
