+++ composeApp/src/commonMain/kotlin/com/crosspaste/signal/PreKeyBundleInterface.kt
@@ -1,0 +1,3 @@
+package com.crosspaste.signal
+
+interface PreKeyBundleInterface
