+++ composeApp/src/commonMain/kotlin/com/crosspaste/signal/PreKeySignalMessageInterface.kt
@@ -1,0 +1,9 @@
+package com.crosspaste.signal
+
+interface PreKeySignalMessageFactory {
+    fun createPreKeySignalMessage(bytes: ByteArray): PreKeySignalMessageInterface
+}
+
+interface PreKeySignalMessageInterface {
+    fun getSignedPreKeyId(): Int
+}
