+++ composeApp/src/commonMain/kotlin/com/crosspaste/signal/SessionBuilderFactory.kt
@@ -1,0 +1,6 @@
+package com.crosspaste.signal
+
+interface SessionBuilderFactory {
+
+    fun createSessionBuilder(signalAddress: SignalAddress): SessionBuilderInterface
+}
