+++ composeApp/src/commonMain/kotlin/com/crosspaste/signal/SessionBuilderInterface.kt
@@ -1,0 +1,6 @@
+package com.crosspaste.signal
+
+interface SessionBuilderInterface {
+
+    fun process(preKeyBundle: PreKeyBundleInterface)
+}
