+++ composeApp/src/desktopMain/kotlin/com/crosspaste/signal/DesktopSessionBuilderFactory.kt
@@ -1,0 +1,9 @@
+package com.crosspaste.signal
+
+class DesktopSessionBuilderFactory(
+    private val signalProtocolStoreInterface: SignalProtocolStoreInterface,
+) : SessionBuilderFactory {
+    override fun createSessionBuilder(signalAddress: SignalAddress): SessionBuilderInterface {
+        return DesktopSessionBuilder(signalProtocolStoreInterface, signalAddress)
+    }
+}
